path fixing
