+++ app/src/pages/Verto/RecoverySeed.vue
@@ -1,174 +1,71 @@
 <template>
   <q-page class="text-black bg-white">
-    <div v-if="step===1" class="standard-content">
-      <h2 class="standard-content--desc"></h2>
-      <div class="standard-content--body">
-        <div class="standard-content--body__img column flex-center">
-          <img src="statics/create_restore_bg.png" alt="">
-        </div>
-      </div>
-      <h2 class="standard-content--title">Do you want to create or restore your 24 word mnemonic secret seed phrase?</h2>
-      <div class="standard-content--footer">
-        <q-btn flat class="action-link restore purple" color="black" text-color="white" label="Restore" />
-        <q-btn flat class="action-link create purple" color="black" text-color="white" label="Create" @click="step=2" />
-      </div>
-    </div>
-    <div v-if="step===2" class="standard-content">
+    <div class="standard-content">
       <h2 class="standard-content--title">This is your 24 - word recovery seed phrase.</h2>
       <h2 class="standard-content--desc">Save these words in the right order in a secure location.Nobody will be able to help if you lose them!</h2>
       <div class="standard-content--body">
         <div class="standard-content--body__mnemonic">
           <h4 class="standard-content--body__mnemonic--title">Mnemonic</h4>
-          <q-input
-            ref="mnemonic"
-            type="textarea"
-            v-model="mnemonic"
-            @focus="$event.target.select()"
-          />
+          <p class="standard-content--body__mnemonic--desc">voluptate velit esse cillum dolore eu fugiat nulla pariatur Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
         </div>
       </div>
       <div class="standard-content--footer">
-         <q-btn flat class="action-link next" color="black" text-color="white" label="Verify" @click="step=3" />
-      </div>
-    </div>
-    <div v-if="step===3" class="standard-content">
-      <h2 class="standard-content--title">Put the words in the right order</h2>
-      <div class="standard-content--body">
-        <words-order :words="mnemonic" />
-      </div>
-      <div class="standard-content--footer">
-         <q-btn flat class="action-link back" color="black" text-color="white" label="Back" />
-         <q-btn flat class="action-link next" color="black" text-color="white" label="Next" />
+         <q-btn flat class="action-link next" color="black" text-color="white" label="Verify" />
       </div>
     </div>
   </q-page>
 </template>
 
 <script>
-const bip39 = require('bip39')
-import HD from '@/util/hdwallet'
-import WordsOrder from '../../components/Verto/WordsOrder'
+import configManager from '@/util/ConfigManager'
+import { version } from '../../../package.json'
+let platformTools = require('../../util/platformTools')
+if (platformTools.default) platformTools = platformTools.default
 
 export default {
-  components: {
-    WordsOrder
-  },
+  components: {},
   data () {
     return {
-      step: 1,
+      text: '',
+      password: '',
       isPwd: true,
-      goodPassword: false,
-      vertoPassword: null,
-      config: null,
-      chip: null,
-      spinnervisible: false,
-      mnemonic: this.$store.state.settings.mnemonic,
-      arrayMnemonic: [],
-      arrayShuffled: [],
-      arrayOrdered: [],
-      arrayShuffleShow: [],
-      arrayTest2: [],
-      arrayTest3: [],
-      master: null,
-      myWallet: null
+      secrectWordsToken: 'voluptate velit esse cillum dolore eu fugiat nulla pariatur Excepteur ',
+      pword: '',
+      minimizedModal: false,
+      message: '',
+      version: {},
+      network: this.$store.state.settings.network,
+      configPath: ''
     }
   },
-  async created () {
-  },
-  async mounted () {
-    console.log('mne', this.mnemonic)
-    if (!this.mnemonic) {
-      this.createMnemonic()
-    } else {
-
-    }
-  },
-  computed: {
-    rightOrder () {
-      if (JSON.stringify(this.arrayMnemonic) === JSON.stringify(this.arrayOrdered)) {
-        return true
-      } else {
-        return false
-      }
-    }
+  mounted () {
+    this.version = version
+    this.setupPlatformPath()
   },
   methods: {
-    chooseMe (word, index, show) {
-      if (show) {
-        this.arrayOrdered.push(word)
-        this.$set(this.arrayShuffleShow, index, show)
-      } else {
-        this.arrayOrdered = this.arrayOrdered.filter(e => e !== word)
-        let unset = this.arrayShuffled.indexOf(word)
-        this.$set(this.arrayShuffleShow, unset, show)
-      }
-
-      this.$set(this.arrayOrdered)
+    async setupPlatformPath () {
+      this.configPath = await platformTools.filePath()
     },
-    async createMnemonic () {
-      console.log('generating mnemonic')
-      this.mnemonic = bip39.generateMnemonic(256)
-      this.arrayMnemonic = this.mnemonic.split(' ')
-      this.arrayShuffled = [...this.arrayMnemonic]
-      this.shuffle(this.arrayShuffled)
-
-      // this.step = 2
+    goChangePassword: function () {
+      this.$router.push({ path: '/change-password' })
     },
-    async saveMnemonic () {
-      if (this.goodPassword) {
-        this.config.mnemonic = this.mnemonic
-        // this.$store.commit('settings/mnemonic', this.mnemonic)
-        await this.$configManager.updateConfig(this.vertoPassword, this.config)
-        const keys = await HD.Wallet('eos')
-        // Need to pass type? // private key gets saved if no pass?
-        const result = await this.$configManager.saveWalletAndKey('HD EOS Key', this.vertoPassword, null, keys.publicKey, keys.privateKey, 'verto', 'mnemonic')
-
-        if (result && result.success) {
-          try {
-            await this.$configManager.backupConfig()
-            if (this.$q.platform.is.android) {
-              this.$q.notify({ color: 'positive', message: 'Config Saved' })
-            }
-          } catch (e) {
-            // TODO: Exception handling
-          }
-          this.vertoPassword = ''
-          this.$q.notify({ color: 'positive', message: 'EOS Keys created' })
-          this.$router.push('wallet')
+    setNetwork: function () {
+      this.$store.dispatch('settings/toggleNetwork', this.network)
+      this.$q.notify({ message: `Network changed to ${this.network}`, color: 'positive' })
+    },
+    async backupConfig () {
+      try {
+        await configManager.backupConfig()
+        if (this.$q.platform.is.android) {
+          this.$q.notify({ color: 'positive', message: 'Config Saved' })
         }
+      } catch (e) {
+        // TODO: Exception handling
       }
-    },
-    async checkVertoPassword () {
-      const results = await this.$configManager.getConfig(this.vertoPassword)
-      if (results.success) {
-        this.goodPassword = true
-        this.config = results.config
-      } else {
-        this.goodPassword = false
-        return false
-      }
-    },
-    shuffle (array) {
-      var currentIndex = array.length, temporaryValue, randomIndex
-
-      // While there remain elements to shuffle...
-      while (currentIndex !== 0) {
-        // Pick a remaining element...
-        randomIndex = Math.floor(Math.random() * currentIndex)
-        currentIndex -= 1
-
-        // And swap it with the current element.
-        temporaryValue = array[currentIndex]
-        array[currentIndex] = array[randomIndex]
-        array[randomIndex] = temporaryValue
-      }
-
-      return array
     }
   }
 }
 </script>
-
 <style lang="scss" scoped>
 @import "~@/assets/styles/variables.scss";
 .standard-content{
@@ -238,9 +135,6 @@ export default {
       }
       &.back{
         background-color: #B0B0B0 !important;
-      }
-      &.purple{
-        background-color: #7900FF !important;
       }
     }
