--- app/src/pages/verto/ChainTools.vue
@@ -1,63 +1,0 @@
-<template>
-  <q-page class="column text-black bg-grey-12" style="padding-bottom: 50px">
-    <profile-header />
-    <chain-tools />
-  </q-page>
-</template>
-
-<script>
-import ChainTools from '../../components/verto/ChainTools'
-import ProfileHeader from '../../components/verto/ProfileHeader'
-import configManager from '@/util/ConfigManager'
-import { version } from '../../../package.json'
-let platformTools = require('../../util/platformTools')
-if (platformTools.default) platformTools = platformTools.default
-
-export default {
-  components: {
-    ChainTools,
-    ProfileHeader
-  },
-  data () {
-    return {
-      pword: '',
-      minimizedModal: false,
-      message: '',
-      version: {},
-      network: this.$store.state.settings.network,
-      configPath: ''
-    }
-  },
-  mounted () {
-    this.version = version
-    this.setupPlatformPath()
-  },
-  methods: {
-    async setupPlatformPath () {
-      this.configPath = await platformTools.filePath()
-    },
-    goChangePassword: function () {
-      this.$router.push({ path: '/change-password' })
-    },
-    setNetwork: function () {
-      this.$store.dispatch('settings/toggleNetwork', this.network)
-      this.$q.notify({ message: `Network changed to ${this.network}`, color: 'positive' })
-    },
-    async backupConfig () {
-      try {
-        await configManager.backupConfig()
-        if (this.$q.platform.is.android) {
-          this.$q.notify({ color: 'positive', message: 'Config Saved' })
-        }
-      } catch (e) {
-        // TODO: Exception handling
-      }
-    }
-  }
-}
-</script>
-<style lang="scss" scoped>
-  /deep/ .profile-wrapper--header{
-    margin-bottom: 0px;
-  }
-</style>
