--- app/src/pages/verto/CreateVertoPassword.vue
@@ -1,151 +1,0 @@
-<template>
-  <q-page class="text-black bg-white">
-    <div class="standard-content">
-      <h2 class="standard-content--title">Create your Verto Password</h2>
-      <h2 class="standard-content--desc">The seed phrase will now be added to your config after confirming the password.</h2>
-      <div class="standard-content--body">
-        <div class="standard-content--body__img column flex-center">
-          <img src="statics/password_bg.png" alt="">
-        </div>
-        <div class="standard-content--body__form">
-          <q-input v-model="password" rounded outlined color="purple" :type="isPwd ? 'password' : 'text'" label="Create Verto Password" hint="*Minimum of 8 characters">
-            <template v-slot:append>
-              <q-icon
-                :name="isPwd ? 'visibility_off' : 'visibility'"
-                class="cursor-pointer"
-                @click="isPwd = !isPwd"
-              />
-            </template>
-          </q-input>
-        </div>
-      </div>
-      <div class="standard-content--footer">
-         <q-btn flat class="action-link next" color="black" text-color="white" label="Next" />
-      </div>
-    </div>
-  </q-page>
-</template>
-
-<script>
-import configManager from '@/util/ConfigManager'
-import { version } from '../../../package.json'
-let platformTools = require('../../util/platformTools')
-if (platformTools.default) platformTools = platformTools.default
-
-export default {
-  components: {},
-  data () {
-    return {
-      text: '',
-      password: '',
-      isPwd: true,
-      secrectWordsToken: 'voluptate velit esse cillum dolore eu fugiat nulla pariatur Excepteur ',
-      pword: '',
-      minimizedModal: false,
-      message: '',
-      version: {},
-      network: this.$store.state.settings.network,
-      configPath: ''
-    }
-  },
-  mounted () {
-    this.version = version
-    this.setupPlatformPath()
-  },
-  methods: {
-    async setupPlatformPath () {
-      this.configPath = await platformTools.filePath()
-    },
-    goChangePassword: function () {
-      this.$router.push({ path: '/change-password' })
-    },
-    setNetwork: function () {
-      this.$store.dispatch('settings/toggleNetwork', this.network)
-      this.$q.notify({ message: `Network changed to ${this.network}`, color: 'positive' })
-    },
-    async backupConfig () {
-      try {
-        await configManager.backupConfig()
-        if (this.$q.platform.is.android) {
-          this.$q.notify({ color: 'positive', message: 'Config Saved' })
-        }
-      } catch (e) {
-        // TODO: Exception handling
-      }
-    }
-  }
-}
-</script>
-<style lang="scss" scoped>
-@import "~@/assets/styles/variables.scss";
-.standard-content{
-  padding: 5% 10%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100vh !important;
-  &--title{
-    font-size: 35px;
-    font-weight: $bold;
-    position: relative;
-    line-height: 50px;
-    font-family: $Titillium;
-    margin-top: 40px;
-    margin-bottom: 40px;
-  }
-  &--desc{
-    margin-top: -20px;
-    margin-bottom: 40px;
-    font-size: 18px;
-    font-weight: $regular;
-    position: relative;
-    line-height: 26px;
-    font-family: $Titillium;
-    color: $mainColor;
-  }
-  &--body{
-    &__img{
-      min-height: 250px;
-      img{
-        max-width: 90%;
-      }
-    }
-    &__form{
-      /deep/ .q-field__native{
-        padding-left: 8px;
-        font-size: 16px;
-        font-weight: $regular;
-      }
-      /deep/ .q-field__label{
-        font-family: $Titillium;
-        font-weight: $bold;
-        font-size: 18px;
-        padding-left: 10px;
-      }
-    }
-  }
-  &--footer{
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: flex-end;
-    min-height: 100px;
-    .action-link{
-      height: 50px;
-      text-transform: initial !important;
-      font-size: 16px;
-      letter-spacing: .5px;
-      border-radius: 40px;
-      width: 110px;
-      margin-left: 10px;
-      &.next{
-        background-color: #7900FF !important;
-      }
-      &.back{
-        background-color: #B0B0B0 !important;
-      }
-    }
-
-  }
-}
-</style>
