--- app/src/pages/verto/Wallets.vue
@@ -1,67 +1,0 @@
-<template>
-  <q-page class="column text-black bg-grey-12" style="padding-bottom: 50px">
-    <profile-header class="marg" />
-    <wallets />
-  </q-page>
-</template>
-
-<script>
-import Wallets from '../../components/verto/Wallets'
-import ProfileHeader from '../../components/verto/ProfileHeader'
-import configManager from '@/util/ConfigManager'
-import { version } from '../../../package.json'
-let platformTools = require('../../util/platformTools')
-if (platformTools.default) platformTools = platformTools.default
-
-export default {
-  components: {
-    ProfileHeader,
-    Wallets
-  },
-  data () {
-    return {
-      pword: '',
-      minimizedModal: false,
-      message: '',
-      version: {},
-      network: this.$store.state.settings.network,
-      configPath: ''
-    }
-  },
-  mounted () {
-    this.version = version
-    this.setupPlatformPath()
-  },
-  methods: {
-    async setupPlatformPath () {
-      this.configPath = await platformTools.filePath()
-    },
-    goChangePassword: function () {
-      this.$router.push({ path: '/change-password' })
-    },
-    setNetwork: function () {
-      this.$store.dispatch('settings/toggleNetwork', this.network)
-      this.$q.notify({ message: `Network changed to ${this.network}`, color: 'positive' })
-    },
-    async backupConfig () {
-      try {
-        await configManager.backupConfig()
-        if (this.$q.platform.is.android) {
-          this.$q.notify({ color: 'positive', message: 'Config Saved' })
-        }
-      } catch (e) {
-        // TODO: Exception handling
-      }
-    }
-  }
-}
-</script>
-<style lang="scss" scoped>
-  .marg{
-    /deep/ .profile-wrapper{
-      &--header{
-        margin-bottom: 0px;
-      }
-    }
-  }
-</style>
